Fix merge conflict & modularize scoop scoring
Resolves a latent merge conflict in how the `state` variable is used.

Modularizes the code used to score and rank scoops so that we can re-use
it for scoring scoops available for other players.

Resolves an issue where it took O(n) time to pop the best flavor off
of the queue. Instead of sorting in reverse order and popping the first
element we can sort in non-decreasing order and pop the last element in
constant time.